Move `InvalidMemoryAccessException` to API
