+++ compiler/src/main/java/org/qbicc/interpreter/InvalidMemoryAccessException.java
@@ -1,4 +1,4 @@
-package org.qbicc.interpreter.impl;
+package org.qbicc.interpreter;
 
 import java.io.Serial;
